pkg lb; added X-Forwarded-Port and improved X- headers
