+++ pkg/lb/httpcommon.go
@@ -94,6 +94,8 @@ func (e *httpError) Unwrap() error {
 
 type httpReqDesc struct {
 	leName                string
+	leHost                string
+	lePort                string
 	leTLS                 bool
 	feName                string
 	feConn                *bufConn
@@ -359,34 +361,4 @@ func groupHTTPStatusMethod(method string) string {
 		groupped = "unknown"
 	}
 	return groupped
-}
-
-func validOptionalPort(port string) bool {
-	if port == "" {
-		return true
-	}
-	if port[0] != ':' {
-		return false
-	}
-	for _, b := range port[1:] {
-		if b < '0' || b > '9' {
-			return false
-		}
-	}
-	return true
-}
-
-func splitHostPort(hostport string) (host, port string) {
-	host = hostport
-
-	colon := strings.LastIndexByte(host, ':')
-	if colon != -1 && validOptionalPort(host[colon:]) {
-		host, port = host[:colon], host[colon+1:]
-	}
-
-	if strings.HasPrefix(host, "[") && strings.HasSuffix(host, "]") {
-		host = host[1 : len(host)-1]
-	}
-
-	return
 }
